+++ include/themes.h
@@ -125,13 +125,13 @@ extern void font_setDefaultTextColors(HDC hdc, THEME_DATA* pTheme);
  * Returns the foreground color of a text style with a given name in the current theme.
  * If the style cannot be found a default value is returned.
  */
-extern long theme_textStyleForeground(char* pszStylename, long lDefault);
+extern long theme_textStyleForeground(const char* pszStylename, long lDefault);
 
 /*
  * Returns the background color of a text style with a given name in the current theme.
  * If the style cannot be found a default value is returned.
  */
-long theme_textStyleBackground(char* pszStylename, long lDefault);
+long theme_textStyleBackground(const char* pszStylename, long lDefault);
 
 #define THEMES_H